+++ src/app/geografia/provincias/province-list.tsx
@@ -1,0 +1,15 @@
+import { IProvinceSchemeAPI, ProvinceItem } from "./province-item";
+
+export const ProvinceList = ({
+	provinces,
+}: {
+	provinces: IProvinceSchemeAPI[];
+}) => {
+	return (
+		<div>
+			{provinces.map((province) => (
+				<ProvinceItem key={province.id} {...province} />
+			))}
+		</div>
+	);
+};
